requirements.txt
vk_api==11.9.1
PyYAML==5.4.1
